Complete refactor

+++ app/src/main/java/hu/krisz768/bettertuke/IncomingBusFragment/BottomSheetIncomingBusFragment.java
@@ -1,5 +1,6 @@
 package hu.krisz768.bettertuke.IncomingBusFragment;
 
+import android.annotation.SuppressLint;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
@@ -21,8 +22,11 @@ import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
 import java.util.Locale;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
 
-import hu.krisz768.bettertuke.Database.BusJaratok;
+import hu.krisz768.bettertuke.Database.BusLine;
 import hu.krisz768.bettertuke.Database.BusPlaces;
 import hu.krisz768.bettertuke.Database.BusStops;
 import hu.krisz768.bettertuke.HelperProvider;
@@ -32,7 +36,7 @@ import hu.krisz768.bettertuke.MainActivity;
 import hu.krisz768.bettertuke.R;
 import hu.krisz768.bettertuke.UserDatabase.UserDatabase;
 import hu.krisz768.bettertuke.api_interface.TukeServerApi;
-import hu.krisz768.bettertuke.api_interface.models.IncommingBusRespModel;
+import hu.krisz768.bettertuke.api_interface.models.IncomingBusRespModel;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -55,8 +59,8 @@ public class BottomSheetIncomingBusFragment extends Fragment {
 
     private IncomingBusStopSelectorAdapter Ibssa;
     private IncomingBusListFragment InBusFragment;
-    private Thread UpdateThread;
-    private volatile boolean UpdateThreadRun = true;
+
+    ScheduledExecutorService UpdateLoop;
 
     public BottomSheetIncomingBusFragment() {
         // Required empty public constructor
@@ -89,29 +93,25 @@ public class BottomSheetIncomingBusFragment extends Fragment {
         // Inflate the layout for this fragment
         View view = inflater.inflate(R.layout.fragment_bottom_sheet_incoming_bus_view, container, false);
 
-        TextView Teszt = view.findViewById(R.id.BusStopName);
+        TextView BusStopName = view.findViewById(R.id.BusStopName);
 
         ImageView ScheduleButton = view.findViewById(R.id.StopScheduleButton);
         ScheduleButton.setImageBitmap(HelperProvider.getBitmap(HelperProvider.Bitmaps.MapStopSelected));
-        ScheduleButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                ((MainActivity)getActivity()).ShowSchedule(mStop, null, null, null, false);
-            }
-        });
-
-        ImageView FavButton = view.findViewById(R.id.StopFavoriteButton);
-        UserDatabase userDatabase = new UserDatabase(getContext());
-
-        if (userDatabase.IsFavorite(UserDatabase.FavoriteType.Stop, Integer.toString(mStop))) {
-            FavButton.setImageBitmap(HelperProvider.getBitmap(HelperProvider.Bitmaps.FaviconOn));
-        } else {
-            FavButton.setImageBitmap(HelperProvider.getBitmap(HelperProvider.Bitmaps.FaviconOff));
+        if (getActivity() != null){
+            ScheduleButton.setOnClickListener(view1 -> ((MainActivity)getActivity()).ShowSchedule(mStop, null, null, null, false));
         }
 
-        FavButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
+        ImageView FavButton = view.findViewById(R.id.StopFavoriteButton);
+        if (getContext() != null){
+            UserDatabase userDatabase = new UserDatabase(getContext());
+
+            if (userDatabase.IsFavorite(UserDatabase.FavoriteType.Stop, Integer.toString(mStop))) {
+                FavButton.setImageBitmap(HelperProvider.getBitmap(HelperProvider.Bitmaps.FaviconOn));
+            } else {
+                FavButton.setImageBitmap(HelperProvider.getBitmap(HelperProvider.Bitmaps.FaviconOff));
+            }
+
+            FavButton.setOnClickListener(view12 -> {
                 if (userDatabase.IsFavorite(UserDatabase.FavoriteType.Stop, Integer.toString(mStop))) {
                     userDatabase.DeleteFavorite(UserDatabase.FavoriteType.Stop, Integer.toString(mStop));
                     FavButton.setImageBitmap(HelperProvider.getBitmap(HelperProvider.Bitmaps.FaviconOff));
@@ -119,19 +119,18 @@ public class BottomSheetIncomingBusFragment extends Fragment {
                     userDatabase.AddFavorite(UserDatabase.FavoriteType.Stop, Integer.toString(mStop));
                     FavButton.setImageBitmap(HelperProvider.getBitmap(HelperProvider.Bitmaps.FaviconOn));
                 }
-            }
-        });
-
+            });
+        }
 
         List<BusStops> SelectedPlaceStops = new ArrayList<>();
 
-        for (int i = 0; i < mPlaceList.length; i++) {
-            if (mPlaceList[i].getId() == mPlace) {
-                Teszt.setText(mPlaceList[i].getName());
+        for (BusPlaces busPlaces : mPlaceList) {
+            if (busPlaces.getId() == mPlace) {
+                BusStopName.setText(busPlaces.getName());
 
-                for (int j = 0; j < mStopList.length; j++) {
-                    if (mStopList[j].getFoldhely() == mPlaceList[i].getId()) {
-                        SelectedPlaceStops.add(mStopList[j]);
+                for (BusStops busStops : mStopList) {
+                    if (busStops.getPlace() == busPlaces.getId()) {
+                        SelectedPlaceStops.add(busStops);
                     }
                 }
             }
@@ -161,40 +160,45 @@ public class BottomSheetIncomingBusFragment extends Fragment {
             }
         };
 
-        int scrollposition = 0;
+        int scrollPosition = 0;
         for (int i = 0; i < SelectedPlaceStopsArray.length; i++) {
             if (SelectedPlaceStopsArray[i].getId() == mStop) {
-                scrollposition = i;
+                scrollPosition = i;
                 break;
             }
         }
 
-        smoothScroller.setTargetPosition(scrollposition);
-        StopSelectorRec.getLayoutManager().startSmoothScroll(smoothScroller);
+        smoothScroller.setTargetPosition(scrollPosition);
+        if (StopSelectorRec.getLayoutManager() != null) {
+            StopSelectorRec.getLayoutManager().startSmoothScroll(smoothScroller);
+        }
+
 
         StopSelectorRec.setNestedScrollingEnabled(false);
 
-        StartNewUpdateThread(view);
+        StartNewUpdateThread();
 
         return view;
     }
 
-    private void StartNewUpdateThread(View view) {
-
-        UpdateThread = new Thread(new Runnable() {
-            @Override
-            public void run() {
-                GetIncommingBusesLoop(view);
+    private void StartNewUpdateThread() {
+        if (UpdateLoop != null) {
+            if (!UpdateLoop.isShutdown()){
+                return;
             }
-        });
-        UpdateThread.start();
+        }
+
+        TukeServerApi serverApi = new TukeServerApi(this.getActivity());
+
+        UpdateLoop = Executors.newScheduledThreadPool(1);
+        UpdateLoop.scheduleAtFixedRate(() -> GetIncomingBuses(serverApi), 0, 10, TimeUnit.SECONDS);
     }
 
     @Override
     public void onPause() {
         super.onPause();
 
-        UpdateThreadRun = false;
+        UpdateLoop.shutdown();
     }
 
     @Override
@@ -202,31 +206,41 @@ public class BottomSheetIncomingBusFragment extends Fragment {
         super.onResume();
 
         ResetList();
-        UpdateThreadRun = true;
+        StartNewUpdateThread();
     }
 
+    @Override
+    public void onStop() {
+        super.onStop();
+
+        UpdateLoop.shutdown();
+    }
+
+    @SuppressLint("NotifyDataSetChanged")
     public void OnStopClick(int Id) {
-        ((MainActivity)getActivity()).ChangeStop(Id);
+        if (getActivity() != null){
+            ((MainActivity)getActivity()).ChangeStop(Id);
+        }
+
         mStop = Id;
         Ibssa.setSelectedStop(mStop);
         Ibssa.notifyDataSetChanged();
         ResetList();
 
-        ImageView FavButton = getView().findViewById(R.id.StopFavoriteButton);
+        if (getView() != null && getContext() != null){
+            ImageView FavButton = getView().findViewById(R.id.StopFavoriteButton);
 
-        UserDatabase userDatabase = new UserDatabase(getContext());
-        if (userDatabase.IsFavorite(UserDatabase.FavoriteType.Stop, Integer.toString(mStop))) {
-            FavButton.setImageBitmap(HelperProvider.getBitmap(HelperProvider.Bitmaps.FaviconOn));
-        } else {
-            FavButton.setImageBitmap(HelperProvider.getBitmap(HelperProvider.Bitmaps.FaviconOff));
+            UserDatabase userDatabase = new UserDatabase(getContext());
+            if (userDatabase.IsFavorite(UserDatabase.FavoriteType.Stop, Integer.toString(mStop))) {
+                FavButton.setImageBitmap(HelperProvider.getBitmap(HelperProvider.Bitmaps.FaviconOn));
+            } else {
+                FavButton.setImageBitmap(HelperProvider.getBitmap(HelperProvider.Bitmaps.FaviconOff));
+            }
         }
 
-        new Thread(new Runnable() {
-            @Override
-            public void run() {
-                TukeServerApi serverApi = new TukeServerApi(getActivity());
-                GetIncommingBuses(serverApi);
-            }
+        new Thread(() -> {
+            TukeServerApi serverApi = new TukeServerApi(getActivity());
+            GetIncomingBuses(serverApi);
         }).start();
 
     }
@@ -238,52 +252,27 @@ public class BottomSheetIncomingBusFragment extends Fragment {
         InBusFragment = null;
     }
 
-    private void GetIncommingBusesLoop(View view) {
-            TukeServerApi serverApi = new TukeServerApi(this.getActivity());
-
-            try {
-                while (true) {
-                    if (getContext() == null)
-                        break;
-                    if (!UpdateThreadRun) {
-                        Thread.sleep(1000);
-                        continue;
-                    }
-                    GetIncommingBuses(serverApi);
-
-                    for (int i = 0; i < 20; i++) {
-                        Thread.sleep(500);
-                        if (!UpdateThreadRun)
-                            break;
-                    }
-                }
-            } catch (Exception e) {
-
-            }
-    }
-
-    private void GetIncommingBuses(TukeServerApi serverApi) {
+    private void GetIncomingBuses(TukeServerApi serverApi) {
         try {
-            final int SendStopid = mStop;
+            final int SendStopId = mStop;
             Log.i("Update", "Updating List");
-            IncommingBusRespModel[] BusList = serverApi.getNextIncommingBuses(mStop);
+            IncomingBusRespModel[] BusList = serverApi.getNextIncomingBuses(mStop);
 
             Date currentTime = Calendar.getInstance().getTime();
             SimpleDateFormat Sdf = new SimpleDateFormat("H", Locale.US);
             SimpleDateFormat Sdf2 = new SimpleDateFormat("m", Locale.US);
 
-            for (int i = 0; i < BusList.length; i++) {
-                BusJaratok Bj = BusJaratok.BusJaratokByJaratid(BusList[i].getJaratid(), getContext());
-                if (Bj.getIndulasOra() < Integer.parseInt(Sdf.format(currentTime)) || (Bj.getIndulasOra() == Integer.parseInt(Sdf.format(currentTime)) && Bj.getIndulasPerc() <= Integer.parseInt(Sdf2.format(currentTime)))) {
-                    BusList[i].setElindult(serverApi.getIsBusHasStarted(BusList[i].getJaratid()));
-                    //Log.e("ASDASDASDASD", Bj.getIndulasOra() + ":" + Bj.getIndulasPerc() + " " + Integer.parseInt(Sdf.format(currentTime))+ ":" + Integer.parseInt(Sdf2.format(currentTime)));
+            for (IncomingBusRespModel incomingBusRespModel : BusList) {
+                BusLine Bj = BusLine.BusLinesByLineId(incomingBusRespModel.getLineId(), getContext());
+                if (Bj.getDepartureHour() < Integer.parseInt(Sdf.format(currentTime)) || (Bj.getDepartureHour() == Integer.parseInt(Sdf.format(currentTime)) && Bj.getDepartureMinute() <= Integer.parseInt(Sdf2.format(currentTime)))) {
+                    incomingBusRespModel.setStarted(serverApi.getIsBusHasStarted(incomingBusRespModel.getLineId()));
                 } else {
-                    BusList[i].setElindult(false);
+                    incomingBusRespModel.setStarted(false);
                 }
 
             }
 
-            if(SendStopid != mStop) {
+            if(SendStopId != mStop) {
                 return;
             }
 
@@ -295,22 +284,21 @@ public class BottomSheetIncomingBusFragment extends Fragment {
                             .replace(R.id.BusListFragment, InBusFragment)
                             .commit();
                 } else {
-                    getActivity().runOnUiThread(new Runnable() {
-                        @Override
-                        public void run() {
+                    if (getActivity() != null) {
+                        getActivity().runOnUiThread(() -> {
                             if (InBusFragment != null) {
                                 InBusFragment.UpdateList(BusList);
                             }
 
-                        }
-                    });
+                        });
+                    }
                 }
             } else {
                 InBusFragment = null;
 
-                InfoFragment Ifragment = InfoFragment.newInstance(getResources().getString(R.string.EmptyList), -1);
+                InfoFragment Fragment = InfoFragment.newInstance(getResources().getString(R.string.EmptyList), -1);
                 getChildFragmentManager().beginTransaction()
-                        .replace(R.id.BusListFragment, Ifragment)
+                        .replace(R.id.BusListFragment, Fragment)
                         .commit();
             }
         } catch (Exception e) {
